refactor: split documentation comparator code into multiple files
- move all comparison logic to doc_ext.dart
- move all parsing logic to doc_parser.dart
- move ApiChange definition to api_change.dart
- keep script logic (arg parsing, cmd output, file content retrieval, etc.) in doc_comparator.dart
- update workflow accordingly
