+++ .github/workflows/pr.yaml
@@ -57,8 +57,8 @@ jobs:
       - name: Run api change type detection
         id: detect_change_type
         run: |
-          API_CHANGELOG=$(dart ./generators/lib/documentation_comparator.dart ./lib/documentation.dart \
-          https://raw.githubusercontent.com/emdgroup-liquid/liquid-flutter/refs/heads/main/lib/documentation.dart -v)
+          API_CHANGELOG=$(dart ./generators/lib/doc_comparator/doc_comparator.dart \
+          HEAD:./lib/documentation.dart main:./lib/documentation.dart -v)
 
           API_CHANGE_TYPE=$(echo "$API_CHANGELOG" | tail -n 1 | awk '{print $NF}')
           echo "API_CHANGE_TYPE=$API_CHANGE_TYPE" >> $GITHUB_ENV
